plugins/versioncontrol: When saving version, show diff of pending changes
When saving a VCS version, present to user not just file status
(modified/added/etc.), but also an actual diff of changes to be committed.
This allows user to see if some unexpected/bad changes may be saved, and
in general, to prepare a better change description.

Currently, this feature is implemented for Git.

Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>