WIP formats/zim_markdown: Parse and dump Markdown for links and images.
Zim markup is also parsed for now, and this is effectively a way to convert
it in existing pages: Zim markup is parsed on read, and then saved as
Markdown.

This is still WIP, as "href" part of these tags is still in adhoc Zim format,
which wouldn't be understood by a generic Markdown tool. This is the next
stage of work.

Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>