zim/gui/pageview: InsertLinkDialog: Make "Link" the default button.
I.e. pressing Enter while dialog is active with cause it to complete. That
means that pressing Ctrl+L, then Enter is enough to quickly insert a link
(if the dialog got it right).

A well-known concept of "default dialog button" requires some magic with
Zim UI widgets, i.e. marking a button as "default" is not enough, one also
need to mark widgets in which pressing Enter will be treated as activating
the default button.

Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>