zim/formats/zim_markdown: Implement native Markdown syntax for checkbox lists.
There's a meaning conflict between Markdown's `[x]`, meaning "done", and
Zim's `[x]` meaning "X-crossed" (represented with checkbox diagonally
crossed with red "x", usual meaning is cancelled). (For reference, "done"
is `[*]` in Zim).

Due to this, it's not a case that Zim checkboxes are just superset of
Markdown. We still support Zim's extended states:

* `[>]` - "migrated" (the same as in Zim)
* `[#]` - "X-crossed" (was `[x]` in Zim)

These extended checkboxes will (should) be rendered literally in other
Markdown tools, so it's up to a user whether to use them and to what
extent. (This patch doesn't change anything on UI side, so all 4 checkbox
states are available via UI).

Change-Id: Ib85d6c80d9e560c436c82355add896621dede101
Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>