zim/notebook/index/tags: Sort pages by (full) name.
Previous upstream code sorted by "sortkey" field, which is obfuscated,
ambiguous encoding of "page base name", which is last component of the
full page name. So for example journal pages like "Journal:2021:03:01"
were sorted based on the last component, leading to completely confusing
overall sort with multiple pages, which potentially could be even more
confusing due to ambigious encoding of the "sortkey" field.

Another ideas would have been to fix the "sortkey" field to be
based off the full page name, but it so happens that this field is
actually not used only as a key for sorting, but is actually also
used to query database, making it all even more messier. So, leaving
"sortkey" as is, and just doing clear sort on the actual page name
seems like better approach (even if it will lead to pages of the same
name but with the different case to be sorted apart - it a well-managed
notebook, such pages should not appear, while again, problems with
sorting happen with the core Journal plugin right away).

Change-Id: Ieff555dbcd4432cffe6e67bc38d80f3f4c7c087c
Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>