zim/notebook/index/pages: list_pages: Don't list placeholders.
Primary motivation of this change is to avoid showing a placeholder journal
page as bold ("having entry") in the calendar widget. It's considered that
in other places where .list_pages() is called the situation would be similar -
placeholder pages are rarely useful to show to users. That's why existing
method is modified, instead of adding new one to filter out placeholders.

Change-Id: Ie33bfe0e7bd8a90ffda2ba8178be3e70169f48e7
Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>