zim/plugins/journal: Update calendar widget when reloading/removing page.
If a page for specific date was created, Ctrl+R will now make corresponding
day in the calendar widget bold. Likewise, if a page was deleted, a day
for it will un-bold.

This is pretty crude implementation, which may be not too efficient. It
also doesn't catch e.g. page saving (instead, reload action is needed).
But at least it works and improves UX with bolded-out days in the
calendar (which before left impression of being very flaky and not to
be trusted).

Change-Id: I850322235a01dd9a3b8c1f65fbc99114857adcc9
Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>